Scaling

+++ python/examples/cc3_5_sf_2a_wrapper.py
@@ -181,7 +181,7 @@ def run_inference(inputs):
 
     #post processing (from pcFM)
     pcfm_out = torch.clamp(pcfm_out, min=0.0) # clamp >= 0
-    counts_t = (pcfm_out[:, :n_layers] + 0.5).to(torch.int32)   # CC3 count calibration
+    counts_t = (pcfm_out[:, :n_layers] + 0.5)   # CC3 count calibration
     counts_t *= 0.64
     counts_t = counts_t.to(torch.int32)
